+++ picker/minimap.lua
@@ -1,13 +1,15 @@
 -------------------------------------------------------------------------------
 --[[Picker Hide Minimap]]--
 -------------------------------------------------------------------------------
+local hide_types = table.arr_to_bool{"logistic-container", "electrical-pole", "roboport"}
+
 local function picker_hide_minimap(event)
     local player = game.players[event.player_index]
-    if not player.opened and player.selected and player.selected.type == "logistic-container" and player.game_view_settings.show_minimap then
+    if not player.opened and player.selected and hide_types[player.selected.type] and player.game_view_settings.show_minimap then
         if player.mod_settings["picker-hide-minimap"].value then
             player.game_view_settings.show_minimap = false
         end
-    elseif not player.game_view_settings.show_minimap and not (player.selected and player.selected.type == "logistic-container") then
+    elseif not player.game_view_settings.show_minimap and not (player.selected and hide_types[player.selected.type]) then
         if player.mod_settings["picker-hide-minimap"].value then
             player.game_view_settings.show_minimap = true
         end